+++ web/src/recent_topics_util.js
@@ -1,9 +1,3 @@
-import $ from "jquery";
-
-import * as compose_state from "./compose_state";
-import * as overlays from "./overlays";
-import * as popovers from "./popovers";
-
 let is_rt_visible = false;
 
 export function set_visible(value) {
@@ -12,18 +6,6 @@ export function set_visible(value) {
 
 export function is_visible() {
     return is_rt_visible;
-}
-
-export function is_in_focus() {
-    // Check if user is focused on
-    // recent topics.
-    return (
-        is_visible() &&
-        !compose_state.composing() &&
-        !popovers.any_active() &&
-        !overlays.is_overlay_or_modal_open() &&
-        !$(".home-page-input").is(":focus")
-    );
 }
 
 export function get_topic_key(stream_id, topic) {
